requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.5
django-filter>=23.3
psycopg2-binary>=2.9.9
django-environ>=0.11.2
Pillow>=10.2.0
gunicorn>=21.2.0
django-cors-headers>=4.3.0
boto3>=1.34.0
django-storages>=1.14.2
stripe>=8.6.0
